requirements.txt
asgiref==3.2.7
autopep8==1.5.2
beautifulsoup4==4.9.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.5
django-colorfield==0.3.0
django-crispy-forms==1.9.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
-e git+https://github.com/gbozee/django-paystack.git@fd2d44767e4c3531add523181eda8001c5ef6170#egg=django_paystack
django-storages==1.9.1
djangorestframework==3.11.0
dropbox==10.1.1
future==0.18.2
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gTTS-token==1.1.3
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.9
Paperboy==1.0.1
Pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
requests-async @ https://github.com/encode/requests-async/archive/master.zip
rfc3986==1.4.0
rsa==4.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
